refactor: 코드 포매터 적용
