+++ app/src/main/java/nextstep/jwp/WebServer.java
@@ -1,12 +1,12 @@
 package nextstep.jwp;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import java.io.IOException;
 import java.net.ServerSocket;
 import java.net.Socket;
 import java.util.stream.Stream;
+import nextstep.jwp.http.RequestHandler;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class WebServer {
 
@@ -18,6 +18,13 @@ public class WebServer {
 
     public WebServer(int port) {
         this.port = checkPort(port);
+    }
+
+    public static int defaultPortIfNull(String[] args) {
+        return Stream.of(args)
+            .findFirst()
+            .map(Integer::parseInt)
+            .orElse(WebServer.DEFAULT_PORT);
     }
 
     public void run() {
@@ -36,13 +43,6 @@ public class WebServer {
         while ((connection = serverSocket.accept()) != null) {
             new Thread(new RequestHandler(connection)).start();
         }
-    }
-
-    public static int defaultPortIfNull(String[] args) {
-        return Stream.of(args)
-                .findFirst()
-                .map(Integer::parseInt)
-                .orElse(WebServer.DEFAULT_PORT);
     }
 
     private int checkPort(int port) {
