+++ app/src/main/java/nextstep/jwp/controller/Controller.java
@@ -1,9 +1,11 @@
 package nextstep.jwp.controller;
 
-import nextstep.jwp.HttpRequest;
-import nextstep.jwp.HttpResponse;
+import nextstep.jwp.http.HttpRequest;
+import nextstep.jwp.http.HttpResponse;
 
 public interface Controller {
+
     HttpResponse get(HttpRequest request);
+
     HttpResponse post(HttpRequest request);
 }
