+++ app/src/main/java/nextstep/jwp/controller/NoController.java
@@ -1,12 +1,7 @@
 package nextstep.jwp.controller;
 
-import java.util.Map;
-import java.util.Objects;
-import nextstep.jwp.HttpRequest;
-import nextstep.jwp.HttpResponse;
-import nextstep.jwp.HttpStatus;
-import nextstep.jwp.db.InMemoryUserRepository;
-import nextstep.jwp.model.User;
+import nextstep.jwp.http.HttpRequest;
+import nextstep.jwp.http.HttpResponse;
 
 public class NoController implements Controller {
