+++ app/src/main/java/nextstep/jwp/http/HttpRequest.java
@@ -1,17 +1,15 @@
-package nextstep.jwp;
+package nextstep.jwp.http;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 public class HttpRequest {
