+++ app/src/main/java/nextstep/jwp/http/HttpResponse.java
@@ -1,4 +1,4 @@
-package nextstep.jwp;
+package nextstep.jwp.http;
 
 import java.util.HashMap;
 import java.util.Map;
