+++ app/src/main/java/nextstep/jwp/http/HttpStatus.java
@@ -1,4 +1,4 @@
-package nextstep.jwp;
+package nextstep.jwp.http;
 
 public enum HttpStatus {
     OK(200, "OK"),
