+++ app/src/main/java/nextstep/jwp/model/User.java
@@ -4,10 +4,10 @@ import java.util.Objects;
 
 public class User {
 
-    private long id;
     private final String account;
     private final String password;
     private final String email;
+    private long id;
 
     public User(long id, String account, String password, String email) {
         this.id = id;
@@ -33,11 +33,11 @@ public class User {
     @Override
     public String toString() {
         return "User{" +
-                "id=" + id +
-                ", account='" + account + '\'' +
-                ", email='" + email + '\'' +
-                ", password='" + password + '\'' +
-                '}';
+            "id=" + id +
+            ", account='" + account + '\'' +
+            ", email='" + email + '\'' +
+            ", password='" + password + '\'' +
+            '}';
     }
 
     public boolean isEmptyId() {
