+++ app/src/test/java/nextstep/jwp/RequestHandlerTest.java
@@ -1,6 +1,6 @@
 package nextstep.jwp;
 
-import java.util.Optional;
+import nextstep.jwp.http.RequestHandler;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
